# DMatrix optimizations