--- a/src/data/data.cc
+++ b/src/data/data.cc
@@ -839,33 +839,36 @@ uint64_t SparsePage::Push(const AdapterBatchT& batch, float missing, int nthread
     return max_columns;
   }
   std::vector<std::vector<uint64_t>> max_columns_vector(nthread);
-
+  dmlc::OMPException exec;
   // First-pass over the batch counting valid elements
 #pragma omp parallel num_threads(nthread)
   {
-    int tid = omp_get_thread_num();
-    size_t begin = tid*thread_size;
-    size_t end = tid != (nthread-1) ? (tid+1)*thread_size : batch_size;
-    max_columns_vector[tid].resize(1, 0);
-    uint64_t& max_columns_local = max_columns_vector[tid][0];
+    exec.Run([&]() {
+      int tid = omp_get_thread_num();
+      size_t begin = tid*thread_size;
+      size_t end = tid != (nthread-1) ? (tid+1)*thread_size : batch_size;
+      max_columns_vector[tid].resize(1, 0);
+      uint64_t& max_columns_local = max_columns_vector[tid][0];
 
-    for (size_t i = begin; i < end; ++i) {
-      auto line = batch.GetLine(i);
-      for (auto j = 0ull; j < line.Size(); j++) {
-        auto element = line.GetElement(j);
-        const size_t key = element.row_idx - base_rowid;
-        CHECK_GE(key,  builder_base_row_offset);
-        max_columns_local =
-            std::max(max_columns_local, static_cast<uint64_t>(element.column_idx + 1));
+      for (size_t i = begin; i < end; ++i) {
+        auto line = batch.GetLine(i);
+        for (auto j = 0ull; j < line.Size(); j++) {
+          auto element = line.GetElement(j);
+          const size_t key = element.row_idx - base_rowid;
+          CHECK_GE(key,  builder_base_row_offset);
+          max_columns_local =
+              std::max(max_columns_local, static_cast<uint64_t>(element.column_idx + 1));
 
-        if (!common::CheckNAN(element.value) && element.value != missing) {
-          // Adapter row index is absolute, here we want it relative to
-          // current page
-          builder.AddBudget(key, tid);
+          if (!common::CheckNAN(element.value) && element.value != missing) {
+            // Adapter row index is absolute, here we want it relative to
+            // current page
+            builder.AddBudget(key, tid);
+          }
         }
       }
-    }
+    });
   }
+  exec.Rethrow();
   for (const auto & max : max_columns_vector) {
     max_columns = std::max(max_columns, max[0]);
   }
@@ -876,20 +879,23 @@ uint64_t SparsePage::Push(const AdapterBatchT& batch, float missing, int nthread
 
 #pragma omp parallel num_threads(nthread)
   {
-    int tid = omp_get_thread_num();
-    size_t begin = tid*thread_size;
-    size_t end = tid != (nthread-1) ? (tid+1)*thread_size : batch_size;
-    for (size_t i = begin; i < end; ++i) {
-      auto line = batch.GetLine(i);
-      for (auto j = 0ull; j < line.Size(); j++) {
-        auto element = line.GetElement(j);
-        const size_t key = (element.row_idx - base_rowid);
-        if (!common::CheckNAN(element.value) && element.value != missing) {
-          builder.Push(key, Entry(element.column_idx, element.value), tid);
+    exec.Run([&]() {
+      int tid = omp_get_thread_num();
+      size_t begin = tid*thread_size;
+      size_t end = tid != (nthread-1) ? (tid+1)*thread_size : batch_size;
+      for (size_t i = begin; i < end; ++i) {
+        auto line = batch.GetLine(i);
+        for (auto j = 0ull; j < line.Size(); j++) {
+          auto element = line.GetElement(j);
+          const size_t key = (element.row_idx - base_rowid);
+          if (!common::CheckNAN(element.value) && element.value != missing) {
+            builder.Push(key, Entry(element.column_idx, element.value), tid);
+          }
         }
       }
-    }
+    });
   }
+  exec.Rethrow();
   omp_set_num_threads(nthread_original);
 
   return max_columns;