Observed https://github.com/dmlc/xgboost/issues/5530 slowdown in DMatrix creation was fixed.

Performance improvements:
Data set | DMatrix create time(master), s | DMatrix create time | Speedup
-- | -- | -- | --
airline-ohe with pandas DF | 13.25 |  8.3 |  1.6
airline-ohe as contiguous numpy| 6.6 |  1.75 |  3.7
higgs1m as contiguous numpy| 0.28 |  0.077 |  3.6

"Dmatrix create + Prediction" improvements (**depth = 1, n_iterations = 10**):
  | airline-ohe with pandas DF | airline-ohe contiguous numpy | higgs1m contiguous numpy
-- | -- | -- | --
Master | 13.37 | 6.8 | 0.296
Dmatrix Opt | 8.44 | 1.86 | 0.085
Speedup | 1.58 | 3.65 | 3.4