@ShvetsKS 2020-07-21
Hi @trivialfis, is it possible to change the default value of `nthread` parameter in DMAtrix initialization?https://github.com/dmlc/xgboost/pull/5877/files#diff-fd53d68e0037d3512896122d1248d969R448

as I understand default parameter `nthreads` for DMAtrix creation was set to  `1` (https://github.com/dmlc/xgboost/blob/master/python-package/xgboost/core.py#L425) as there was a negative scaling with increasing number of threads in `uint64_t SparsePage::Push(const AdapterBatchT& batch, float missing, int nthread)` method? 
It was changed to `-1` as currently we can see speed up with  threading.

Thanks you in advance.

fix of problem with omp max reduction on VS is in progress.