@ShvetsKS 2020-07-23
Not sure that failed step "continuous-integration/travis-ci/pr" is related to current changes.
@trivialfis , @hcho3  sorry for bothering you, should I rerun it?