@ShvetsKS 2020-08-12
@trivialfis could you take a look at failed tests, I'm not sure that it's related to this PR.
there is a problem with `DeviceQuantileDMatrix `, and fails: `Multiple processes within communication group running on same CUDA device is not supported`. After I look at the code, seems current changes should not affect gpu implementation.

and I can hardly believe that 'Error in loadNamespace(name) : there is no package called 'DiagrammeR' ' in r test is related to current changes.