@ShvetsKS 2020-08-13
> Let me take a look later. This is new.

in another PR(temporary) with the same changes: https://github.com/dmlc/xgboost/pull/6007 step with r tests (test-r-3.5.3) have been passed.