@ShvetsKS 2020-08-17
@trivialfis could you approve this PR, as seems there is no real problem with testing?