@ShvetsKS 2020-08-18
> LGTM overall. Could you please help addressing some minor issues?

@trivialfis  thanks for review