@ShvetsKS 2020-08-19
@trivialfis is failed `codecov/patch` check critical for merging?